cleanup in EditHelper #98
